Refactor acceptance tests
* Process.spawn with :pgroup => true option. Otherwise foreman was
  unable to kill its subprocesses

* Split into multiple files
