+++ test/test_helper.rb
@@ -2,4 +2,4 @@ require 'minitest/autorun'
 require 'mocha'
 require 'm2r'
 
-Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
+require 'support/test_handler'
